src/services/processor/requirements.txt
pika==1.3.2
psycopg2-binary==2.9.9
jsonschema==4.20.0
prometheus-client==0.19.0

